749 extract test helper; calibrate
